Rewrite
Change icon color & added text to string & action on button
