+++ README.md
@@ -3,7 +3,5 @@
 - apk
 https://disk.yandex.ru/d/RYYxnTQxQ4y57Q
 - preview
-![Alt Text](./Screen_Recording_20231103-100151_LABA1_compose.gif)
+![](./Screen_Recording_20231103-100151_LABA1_compose.gif)
 
-
-[![Watch the video]](https://github.com/MikeTsvirkunov/DC_Laba1_compose/raw/dev/Screen_Recording_20231103-100151_LABA1_compose.mp4)
